예외 처리 로직 추가

+++ src/main/java/com/f5/onepageresumebe/domain/service/PortfolioService.java
@@ -5,15 +5,13 @@ import com.f5.onepageresumebe.domain.repository.*;
 
 import com.f5.onepageresumebe.domain.repository.querydsl.PortfolioQueryRepository;
 import com.f5.onepageresumebe.domain.repository.querydsl.UserQueryRepository;
+import com.f5.onepageresumebe.exception.ErrorCode;
 import com.f5.onepageresumebe.exception.customException.CustomAuthenticationException;
+import com.f5.onepageresumebe.exception.customException.CustomException;
 import com.f5.onepageresumebe.security.SecurityUtil;
 import com.f5.onepageresumebe.util.PorfUtil;
 import com.f5.onepageresumebe.util.ProjectUtil;
 import com.f5.onepageresumebe.util.StackUtil;
-import com.f5.onepageresumebe.web.dto.career.requestDto.CareerRequestDto;
-import com.f5.onepageresumebe.web.dto.career.requestDto.CareerListRequestDto;
-import com.f5.onepageresumebe.web.dto.career.responseDto.CareerListResponseDto;
-import com.f5.onepageresumebe.web.dto.career.responseDto.CareerResponseDto;
 import com.f5.onepageresumebe.web.dto.porf.ChangeStatusDto;
 import com.f5.onepageresumebe.web.dto.porf.requestDto.PorfIntroRequestDto;
 import com.f5.onepageresumebe.web.dto.porf.requestDto.PorfProjectRequestDto;
@@ -21,7 +19,6 @@ import com.f5.onepageresumebe.web.dto.porf.requestDto.PorfTemplateRequestDto;
 import com.f5.onepageresumebe.web.dto.porf.responseDto.PorfIntroResponseDto;
 import com.f5.onepageresumebe.web.dto.porf.responseDto.PorfResponseDto;
 import com.f5.onepageresumebe.web.dto.project.responseDto.ProjectDetailListResponseDto;
-import com.f5.onepageresumebe.web.dto.project.responseDto.ProjectDetailResponseDto;
 import com.f5.onepageresumebe.web.dto.stack.StackDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -33,6 +30,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.f5.onepageresumebe.exception.ErrorCode.INVALID_INPUT_ERROR;
+
 
 @RequiredArgsConstructor //롬북을 통해서 간단하게 생성자 주입 방식의 어노테이션으로 fjnal이 붙거나 @notNull이 붙은 생성자들을 자동 생성해준다.
 @Service
@@ -42,7 +41,6 @@ public class PortfolioService {
     private final PortfolioRepository portfolioRepository;
     private final StackRepository stackRepository;
     private final PortfolioStackRepository portfolioStackRepository;
-    private final UserStackRepository userStackRepository;
     private final CareerRepository careerRepository;
     private final ProjectRepository projectRepository;
     private final ProjectImgRepository projectImgRepository;
@@ -86,10 +84,15 @@ public class PortfolioService {
         Portfolio portfolio = portfolioQueryRepository.findByUserEmailFetchUser(userEmail).orElseThrow(
                 () -> new IllegalArgumentException("포트폴리오가 존재하지 않습니다"));
 
+        List<String> stacks = requestDto.getStack();
+        if (stacks.size()<3){
+            throw new CustomException("포트폴리오 스택을 3개 이상 선택해 주세요.",INVALID_INPUT_ERROR);
+        }
+
         //기존에 있는 스택 모두 삭제
         portfolioStackRepository.deleteAllByPorfId(portfolio.getId());
 
-        insertStacksInPortfolio(portfolio, requestDto.getStack());
+        insertStacksInPortfolio(portfolio, stacks);
     }
 
     @Transactional
@@ -100,7 +103,12 @@ public class PortfolioService {
         Portfolio portfolio = portfolioQueryRepository.findByUserEmailFetchUser(userEmail).orElseThrow(() ->
                 new IllegalArgumentException("포트폴리오가 존재하지 않습니다"));
 
-        String changedStatus = portfolio.changeStatus(dto.getStatus());
+        String status = dto.getStatus();
+        if(!("public".equals(status) || "private".equals(status)) ){
+            throw new CustomException("포트폴리오 상태값은 public 이거나 private 입니다.", INVALID_INPUT_ERROR);
+        }
+
+        String changedStatus = portfolio.changeStatus(status);
 
         return ChangeStatusDto.builder()
                 .status(changedStatus)
@@ -114,6 +122,10 @@ public class PortfolioService {
                 new IllegalArgumentException("포트폴리오가 존재하지 않습니다"));
 
         List<Integer> projectIds = requestDto.getProjectId();
+
+        if (projectIds.isEmpty()){
+            throw new CustomException("최소 하나의 프로젝트를 선택해 주세요.",INVALID_INPUT_ERROR);
+        }
 
         List<Project> projects = projectRepository.findAllByIds(projectIds);
 
@@ -132,6 +144,10 @@ public class PortfolioService {
                 new IllegalArgumentException("포트폴리오가 존재하지 않습니다"));
 
         List<Integer> projectIds = requestDto.getProjectId();
+
+        if (projectIds.isEmpty()){
+            throw new CustomException("최소 하나의 프로젝트를 선택해 주세요.",INVALID_INPUT_ERROR);
+        }
 
         List<Project> projects = projectRepository.findAllByIds(projectIds);
